refactor: move Swagger Setup to configs folder
